# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RequestStatus = typeof RequestStatus[keyof typeof RequestStatus];


export const RequestStatus = {
  queued: 'queued',
  in_progress: 'in_progress',
  ready: 'ready',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  teacher: 'teacher',
  printer: 'printer',
} as const;

export interface Branch {
  id: number;
  name: string;
}

export interface BranchInput {
  /**
     * @minLength 2
     * @maxLength 80
     */
  name: string;
}

export interface UserProfile {
  clerkId: string;
  fullName: string;
  email: string;
  role: UserRole;
  /** @nullable */
  branchId: number | null;
  /** @nullable */
  branchName: string | null;
  isActive: boolean;
}

export interface PrinterLoginInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  username: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  password: string;
}

export interface SchoolUser {
  clerkId: string;
  fullName: string;
  email: string;
  role: UserRole;
  /** @nullable */
  branchId: number | null;
  /** @nullable */
  branchName: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface UserAssignment {
  role: UserRole;
  /** @nullable */
  branchId: number | null;
  isActive: boolean;
}

export interface DashboardSummary {
  openCount: number;
  urgentCount: number;
  inProgressCount: number;
  readyCount: number;
  completedToday: number;
  overdueCount: number;
}

export interface AttachmentInput {
  /** @minLength 1 */
  objectPath: string;
  /**
     * @minLength 1
     * @maxLength 255
     */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
  /** @minLength 20 */
  ownerToken: string;
}

export interface AttachmentInfo {
  fileIndex: number;
  name: string;
  size: number;
  contentType: string;
}

export interface PrintRequest {
  id: number;
  title: string;
  copies: number;
  dueAt: string;
  /** @nullable */
  note: string | null;
  status: RequestStatus;
  createdAt: string;
  branchId: number;
  branchName: string;
  requesterName: string;
  attachments: AttachmentInfo[];
}

export interface PrintRequestInput {
  /**
     * @minLength 2
     * @maxLength 160
     */
  title: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  requesterName: string;
  /** @minimum 1 */
  branchId: number;
  /**
     * @minimum 1
     * @maximum 5000
     */
  copies: number;
  dueAt: string;
  /**
     * @maxLength 1000
     * @nullable
     */
  note?: string | null;
  /**
     * @minItems 1
     * @maxItems 8
     */
  attachments: AttachmentInput[];
}

export interface RequestStatusInput {
  status: RequestStatus;
}

export interface UploadUrlInput {
  /**
     * @minLength 1
     * @maxLength 255
     */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrl {
  uploadURL: string;
  objectPath: string;
  ownerToken: string;
}

export type PrinterLogout200 = {
  success: boolean;
};

export type ListPrintRequestsParams = {
status?: RequestStatus;
};