# lib/api-zod/src/generated/types/attachmentInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AttachmentInput {
  /** @minLength 1 */
  objectPath: string;
  /**
     * @minLength 1
     * @maxLength 255
     */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
  /** @minLength 20 */
  ownerToken: string;
}