# lib/api-zod/src/generated/types/printRequestInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AttachmentInput } from './attachmentInput';

export interface PrintRequestInput {
  /**
     * @minLength 2
     * @maxLength 160
     */
  title: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  requesterName: string;
  /** @minimum 1 */
  branchId: number;
  /**
     * @minimum 1
     * @maximum 5000
     */
  copies: number;
  dueAt: Date;
  /**
     * @maxLength 1000
     * @nullable
     */
  note?: string | null;
  /**
     * @minItems 1
     * @maxItems 8
     */
  attachments: AttachmentInput[];
}